navbar appears but is yet to be styled or properly function
