+++ Landing_page/style.css
@@ -56,4 +56,6 @@ h3 {
     -webkit-background-clip: text;
     background-clip: text;
   }
+
+
 }
