+++ src/pages/authentication/index.jsx
@@ -1,93 +1,13 @@
 import { auth, provider } from "../../config/firebase-config";
 import { signInWithPopup } from "firebase/auth";
-import { useNavigate } from "react-router-dom";
+import { Route, Routes, useNavigate } from "react-router-dom";
+import { NavBar } from "../authentication/component/NavBar";
+import { Inspiration } from "../authentication/component/navpages/Inspiration";
+import { Features } from "../authentication/component/navpages/Features";
+import { AboutUs } from "../authentication/component/navpages/About_us";
+
+// import { Link } from "react-router-dom";
 import "./styles.css";
-import headshot from "./Gabriels_profesional_shots.jpeg";
-import headshot2 from "./Natalia_profesional_shot.jpeg";
-
-const Header = () => {
-    return (
-    <header className="header">
-        <nav>
-            <table>
-                <tr>
-                    <td><span className="fancy">SPENDY</span></td>
-                    <td><a href="#inspiration" className="btn">What inspires us</a></td>
-                    <td><a href="#features" className="btn">Features</a></td>
-                    <td><a href="#about-us" className="btn">About us</a></td>
-                </tr>
-            </table>
-        </nav>
-    </header>
-    );
-};
-
-const Inspiration = () => {
-    return (
-        <div className="inspiration" id="inspiration"> 
-            <h2><span class="fancy">What inspires us</span></h2>
-            <p>
-            Our journey began as two employees of Banco Popular, selected for a unique opportunity at Holberton Coding School, fully sponsored to innovate in project-based learning. Inspired by Banco Popular's MiBanco app, we recognized its potential for enhancement. Our goal was clear: to reimagine MiBanco, making it as user friendly as possible but also injecting elements of gamification to incentivize user engagement. The result is a fusion of our banking expertise and coding skills, delivering a more intuitive and, rewarding experience for MiBanco users.
-            </p>
-        </div>
-    );
-}
-
-const Features = () => {
-    return (
-        <div className="features" id="features">
-            <th><span class="fancy">Features</span></th>
-            <table>
-                <tr>
-                    <td>
-                        <h3>Easy Tracking</h3>
-                        <p>Our app makes expense tracking effortless. With user-friendly interfaces and streamlined workflows, you can easily input, categorize, and analyze your expenses in just a few clicks. Say goodbye to complicated spreadsheets and clunky interfaces, and hello to simplicity and efficiency.</p>
-                    </td>
-                    <td>
-                        <h3>Intuitive Visuals</h3>
-                        <p>Experience your financial data in a whole new light with our intuitive visuals. Our app offers a single, sleek graph that displays your income and expenses by category, providing a clear overview of your financial landscape. With just a glance, you can see where your money is going and make informed and fast decisions about your spending.</p>
-                    </td>
-                    <td>
-                        <h3>Achievements</h3>
-                        <p>Stay motivated on your financial journey with our achievements feature. Track your progress as you achieve financial milestones, from sticking to your budget to increasing your savings. Our app celebrates your successes and encourages you to keep striving for financial stability and success. With achievable goals and meaningful rewards, you'll feel empowered to take control of your finances and achieve your financial dreams.</p>
-                    </td>
-                </tr>
-            </table>
-        </div>
-    );
-}
-
-
-const AboutUs = () => {
-    return (
-        <div className="about-us" id="about-us">
-            <h2><span class="fancy">About us</span></h2>
-            <p>
-            We're Gabriel Castro and Natalia Rivera, two aspiring developers excited to embark on this journey in the world of tech.
-            As newcomers to the realm of software development, we're eager to learn, grow, and make a positive impact through our work. With a passion for problem-solving and a commitment to continuous improvement, we're diving headfirst into the world of coding, ready to tackle any challenge that comes our way.
-            </p>
-            <p>
-            we're always open to feedback, suggestions, and new connections. If you'd like to learn more about us or connect with us further, feel free to check out our GitHub and LinkedIn profiles:
-            </p>
-            <div>
-            <img src={headshot} alt="headshot" className="headshot"/>
-            <img src={headshot2} alt="headshot" className="headshot"/>
-            </div>
-            <div>
-                <a href="https://www.linkedin.com/in/gabrielcastroflores/" class="button-linkedin" >Gabriel's LinkedIn</a>
-                <a href="https://www.linkedin.com/in/nataliagrivera/" class="button-linkedin">Natalia's LinkedIn</a>
-            </div>
-            <div className="github links">
-                <a href="https://github.com/ThatFireBoi" class="button-git">Gabriel's GitHub</a>
-                <a href="https://github.com/nataliagrivera" class="button-git">Natalia's GitHub</a>
-            </div>
-        </div>
-    );
-}
-
-const Footer = () => {
-    return <footer className="Footer">This is the footer</footer>;
-};
 
 export const Auth = () => {
 
@@ -106,9 +26,11 @@ export const Auth = () => {
     };
     return (
         <div className="landing-page">
-        <div>
-            <Header />
-        </div>
+        <section>
+            <NavBar />
+            <Routes path="/about" element={<AboutUs />}></Routes>
+            <Routes path="/about" element={<AboutUs />}></Routes>
+        </section>
         <body>
         <section>
             <article>
