+++ src/pages/authentication/styles.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400;500;600;700&display=swap');
 
 .user-login-page {
-  width: 100vw;
+  width: 100%;
   height: 100vw;
   display: flex;
   flex-direction: column;
@@ -11,17 +11,36 @@
   text-shadow: -1px 2px 6px rgba(2, 48, 71, 0.54);
   background: rgb(0, 204, 255);
   background: url('./sp.jpg') no-repeat;
+  background-attachment: fixed;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+  }
+
+.user-login-page
+  section {
+    background-attachment: fixed;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+  }  
+
+.navbar {
+  z-index: 1;
+  position: fixed;
+  background-color: #212f4a;
+  width: 100%;
+}
+
+.navbar section {
 }
 
 section {
-  scroll-snap-align: start;
-  scroll-snap-stop: always;
-  scroll-snap-type: y mandatory;
   display: flex;
   align-items: center;
   justify-content: center;
-  height: calc(100vh);
-  width: 100vw;
+  height: 100vh;
+  width: 100%;
 }
 
 .user-login-page p {
@@ -110,20 +129,14 @@ section {
     background-clip: text;
   }
 }
+.w3-bar-item {
+  padding: 1rem;
+  text-decoration: none;
+  color: #333;
+}
 
-.header {
-  display: flex;
-  justify-content: space-evenly;
-  margin:auto;
-  background-color: #212f4a;
-  font-size: 2.0rem;
-  font-weight: 700;
-  font-family: "Quicksand", sans-serif;
-  height: 80px;
-  position: fixed;
-  width: 100%;
-  padding: 10px; /* Added 'px' unit */
-  overflow: hidden;
+.w3-bar-item:hover {
+  background-color: #ddd;
 }
 
 
@@ -206,9 +219,42 @@ body {
   font-size: 20px;
 }
 
-.features td {
+.inspiration {
+  display: flex;
+  height: 100%;
+  flex-direction: column;
+  background-color: #212f4a;
+  border: 2px solid #212f4a;
   padding: 20px;
-  border: 50px solid #212f4a;
+  border-radius: 20px;
+  margin-bottom: 20px;
+}
+
+.inspiration h2 {
+  font-size: 2.5rem;
+  color: whitesmoke;
+  text-align: center;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
+  Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+}
+
+.inspiration p {
+  font-size: 1.0rem;
+  color: whitesmoke;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
+  Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+}
+
+.features section {
+  background-attachment: fixed;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;;
+}
+
+.features td {
+  padding: 10px;
+  border: 25px solid #212f4a;
   text-align: center;
   transition: transform 0.3s ease-in-out;
   background-color: #212f4a;
@@ -279,31 +325,6 @@ body {
   Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
 }
 
-.inspiration {
-  display: flex;
-  
-  flex-direction: column;
-  background-color: #212f4a;
-  border: 2px solid #212f4a;
-  padding: 20px;
-  border-radius: 20px;
-  margin-bottom: 20px;
-}
-
-.inspiration h2 {
-  font-size: 2.5rem;
-  color: whitesmoke;
-  text-align: center;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-  Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-}
-
-.inspiration p {
-  font-size: 1.0rem;
-  color: whitesmoke;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-  Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-}
 
 
 .headshot {
